Handle fixed form Fortran continuation lines
Fixed form Fortran (aka. 66/77) has continuation lines specified in
column 6 of each lines by putting a character that is not ' ' or '0'
these lines are supposed to be interpreted as part of the statement
above. Since fixed form is not handled explicitly in current version of
language-fortran, preprocessor takes these continuation lines and
appends it to the end of the previous line before sending it to lexer.